feat: changes in samples as per comments,renaming tags, merging numeric samples
