+++ samples/pg-dml-partitioned.js
@@ -23,7 +23,7 @@ function main(
   databaseId = 'my-database',
   projectId = 'my-project-id'
 ) {
-  // [START spanner_pg_partitioned_dml]
+  // [START spanner_postgresql_partitioned_dml]
   /**
    * TODO(developer): Uncomment these variables before running the sample.
    */
@@ -61,7 +61,7 @@ function main(
     }
   }
   pgPartitionedDml();
-  // [END spanner_pg_partitioned_dml]
+  // [END spanner_postgresql_partitioned_dml]
 }
 
 process.on('unhandledRejection', err => {
